Hand cards placement working + card data serialization
Now some of the card data are serialized in JSON and loaded at runtime. Ultimately, card effects will be serialized too.
Now cards in hand are properly placed on the screen.
Now cars in hand are randomly chosen from the player deck at the beginning of the turn.
Now cards are put into the graveyard when played and shuffled back into the deck at the end of the turn.w