Big architecture refactoring
I broke a lot of stuff (most notably the json serialization but anyway its destiny was to be rewritten) but now the core is a referenced DLL and the game files now only contain specific game-related stuff.
UI rework in progress.
Game main system in progress.
Duels with other players is expected to be working soon.